refactor BorrowingMethod component
