+++ components/exercises/chapter-6-oscar/BorrowingMethod/BorrowingMethod.tsx
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import { checkIfExist } from './checkIfExist';
 
 const foodList = {
   hotdog: true,
@@ -6,16 +7,6 @@ const foodList = {
   hamburger: true,
   hasOwnProperty: true,
 };
-
-/**
- * Checks if an object has a specific key.
- * @param list - The object to check.
- * @param key - The key to check for.
- * @returns A boolean indicating whether the object has the specified key.
- */
-export function checkIfExist(list: object, key: string) {
-  return Object.prototype.hasOwnProperty.call(list, key);
-}
 
 const keysArray: string[] = [];
